Latest dmd-play as a service scripts for batocera v41+ users

+++ userdata/system/configs/emulationstation/scripts/game-selected/dmd-simulator.sh
@@ -18,39 +18,49 @@ currSystem="$1"
 currGame="$2"
 forced="$4"
 
-# Only proceed to change marquee if the user has settled on the system and there isn't another event that has superceeded this instance
-source /usr/bin/dmd-helper/es-event-proceed.sh 'game-selected' "${currSystem}" "${currGame}" "${forced}"
+# Only proceed to change marquee if the user has settled on the game and there isn't another event that has superceeded this instance
+source dmd-event-proceed.sh 'game-selected' "${currSystem}" "${currGame}" "${forced}"
 if [ $? -ne 0 ]; then exit; fi
 
 # Ok - safe to proceed to change the game Marquee!
-echo "${progname}CHANGING MARQUEE" >> ${esScriptLogFile}
+debug "CHANGING GAME MARQUEE for: ${currGame}"
 
 foundImage=0
 # Find a custom marque GIF or PNG in /userdata/system/dmd/games/<system>, stripping out special characters
 # lowercase and remove any but a-z and 0-9, remove things in parenthesis
+GAMEBASE=$(basename "${currGame}")
 ROMBASE=$(basename "${currGame}" | sed -e s+"\.[^\.]*$"++)
-ROMMIN=$(echo "${ROMBASE}" | sed -e s+"([^)]*)"+""+g -e s+"[^A-Za-z0-9]"+""+g | tr A-Z a-z) 
-for EXT in gif png
+ROMMIN=$(echo "${ROMBASE}" | sed -e s+"([^)]*)"+""+g -e s+"[^_A-Za-z0-9]"+""+g | tr A-Z a-z) 
+for BASE in "/userdata/system/dmd/games/${currSystem}/${ROMMIN}" "/userdata/roms/${currSystem}/images/${ROMBASE}-marquee" "/userdata/roms/${currSystem}/images/${ROMBASE}-image" 
 do
-    CUS="/userdata/system/dmd/games/${currSystem}/${ROMMIN}.${EXT}"
+	for EXT in gif png
+	do
+		CUS=${BASE}.${EXT}
+		debug "Looking for custom marquee image: ${CUS}"
 	
-	echo "${progname}Looking for custom marquee image: ${CUS}" >> ${esScriptLogFile}
-	
-    if test -e "${CUS}"
-    then
-		echo "${progname}Found custom marquee image, playing via dmd-play" >> ${esScriptLogFile}		
-		foundImage=1
-		source es-dmd-play.sh "${CUS}"
+		if test -e "${CUS}"
+		then
+			debug "Found custom marquee image: ${CUS}, playing via dmd-play"
+			foundImage=1
+			source dmd-play-if-changed.sh "${CUS}"
+			break
+		fi
+	done
+	if [ $foundImage -eq 1 ]
+	then
 		break
-    fi
+	fi
 done
 
-# Finally remove the mutex
-echo "${progname}Desired game is: ${currGame}.  Removing mutex" >> ${esScriptLogFile}
-rmdir /tmp/dmd-mutex > /dev/null 2>&1
+if [ -z ${playAsService} ]
+then
+	# If running interactively, remove the  mutex
+	debug "Game change: ${currGame}, finished.  Removing mutex"
+	rmdir /tmp/dmd-mutex > /dev/null 2>&1
+fi
 
 if [ $foundImage -eq 0 ]
 then
-	echo "${progname}Couldn't find custom nor ES sourced marquee, just display the system marquee for: ${currSystem}" >> ${esScriptLogFile}
-	source /userdata/system/configs/emulationstation/scripts/system-selected/dmd-simulator.sh ${currSystem}
+	debug "Couldn't find custom nor ES sourced marquee, just display the system marquee for: ${currSystem}"
+	source "${esScriptDir}"/system-selected/dmd-simulator.sh ${currSystem}
 fi
